requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.1
langchain>=0.1.13
langchain-core==0.1.42
langchain-groq>=0.1.5
langchain-community==0.0.27
groq>=0.5.0
faiss-cpu
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
aiohttp==3.9.1
textblob==0.17.1
sentence-transformers>=2.2.2
web3==6.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dateutil==2.9.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.28.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
